requirements.txt
asgiref==3.6.0
Django==4.1.5
djangorestframework==3.14.0
joblib==1.2.0
mysqlclient==2.1.1
numpy==1.24.1
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
